backend/app/utils/gemini_client.py
import google.generativeai as genai
import os
from dotenv import load_dotenv

load_dotenv()
genai.configure(api_key=os.getenv("GOOGLE_API_KEY"))

def get_gemini_response(prompt: str):
    try:
        model = genai.GenerativeModel("gemini-2.5-pro")  # or gemini-1.5-pro
        response = model.generate_content(prompt)
        return response.text.strip()
    except Exception as e:
        return f"Error: {str(e)}"
