backend/requirements.txt
fastapi
uvicorn
pymongo
librosa
pydub
numpy
langchain
langchain-google-genai
google-generativeai
python-dotenv
SpeechRecognition
pyaudio
torch
torchvision
torchaudio
transformers
sqlalchemy